Use CBOR maps in signed exchanges.
Instead of arrays with alternating keys and values, which side-steps the
disagreement about CBOR canonical map ordering rules, I now specify one
of the ordering rules and use maps where they make sense.

The one place I'm still not using a map is in the top level of the
exchange representation, where the request/response/payload ordering
seems useful and where we might someday want to add, for example, the
request payload between the request and response.