Try all PEM blocks in private key file.
Change gen-signedexchange to try parsing all blocks in the key file as a
private key. On OpenSSL 1.1.0h, the instructions at README.md result in
a key file that has two blocks, where the first is an EC PARAMETERS
block and the second is the EC PRIVATE KEY. This allows that file to be
read.