Introduce scopes attribute for subresource Web Bundles link.
Currently the explainer of "subresource loading with Web Bundles" is
using a new "scope" attribute of the link element. However the new
attribute should be "scopes" that accept multiple scopes, giving web
developers more flexibility. For example, in this case, some dynamic
contents can be under "https://example.com/dir/api/".

So this PR changes the explainer to use "scopes".