Remove the manifest and critical sections
The idea of this PR is that the "core" Web Bundles specification does not include the manifest or critical sections, but that these sections can be defined in other documents. The motivation is to factor the Web Bundles specification to include just the minimal amount needed to support cases like subresource loading. Subresource loading does not need a manifest. For the critical section, if it is missing, the behavior is likely to be to not require any sections, so it can also be defined and interpreted just where it is needed for more detailed bundle types.