Use uuid-in-package: scheme for opaque origin
The "urn:" scheme is included in the "safelisted schemes" of the HTML
spec, meaning that web sites can register a custom protocol handler that
handles the "urn:" scheme.

To avoid the potential for conflict, this introduces a new URL scheme
"uuid-in-package:" that can be used in resource URLs in web bundles.