Fix "Update Editor's Copy" workflow
It was failing because bikeshed was not installed in the
`martinthomson/i-d-template@v1` docker image. Installing bikeshed in it
is not trivial (that needs to build native extensions but the docker
image does not have gcc), so this patch lets bikeshed run outside the
docker environment.

This also makes this action publish `subresource-loading.html`.