Enable PR preview for subresource loading spec
This enables PR preview for future changes in the subresource loading
spec.

Unfortunately pr-preview does not support multiple specifications in
single repo, so this means we will not get PR previews for loading.bs
after this change.

I think this is beneficial, because we anticipate many changes to
subresource-loading.bs for some time to come, while loading.bs is pretty
stable these days.