Use corsAttributeState of Anonymous in alternate SignedExchange prefetch
This fixes #790.

This changes the corsAttributeState used to create a request for
alternate SignedExchange prefetch from No CORS to Anonymous. See #790
for the motivation of this change.

For consistency this also changes the corsAttributeState for
requestForMatch, but it doesn't make any behavioral changes because the
"matches the stored exchange" algorithm does not refer the request's
mode and credentials mode.